squashed commit from icewind1991 - see #4023
* cleaner checksum check in test
* cleanup checksum wrappers
* fix error when trying to remove non existing entry from cache

+++ lib/private/Files/Cache/Cache.php
@@ -439,6 +439,9 @@ class Cache implements ICache {
 	 */
 	public function remove($file) {
 		$entry = $this->get($file);
+		if (!$entry || !$entry['fileid']) {
+			return;
+		}
 		$sql = 'DELETE FROM `*PREFIX*filecache` WHERE `fileid` = ?';
 		$this->connection->executeQuery($sql, array($entry['fileid']));
 		if ($entry['mimetype'] === 'httpd/unix-directory') {
